+++ frontend/src/pages/MyTemplatesPage/components/TemplateDeleteSelection/TemplateDeleteSelection.tsx
@@ -1,6 +1,5 @@
 import { Button, Flex } from '@/components';
-
-import { ConfirmDeleteModal } from '../';
+import { ConfirmDeleteModal } from '@/pages/MyTemplatesPage/components';
 
 interface Props {
   isEditMode: boolean;
