+++ frontend/src/pages/MyTemplatesPage/components/index.ts
@@ -1,14 +1,14 @@
-export { default as CategoryEditModal } from './CategoryEditModal/CategoryEditModal';
-export { default as CategoryFilterMenu } from './CategoryFilterMenu/CategoryFilterMenu';
-export { default as TagFilterMenu } from './TagFilterMenu/TagFilterMenu';
-export { default as TemplateGrid } from './TemplateGrid/TemplateGrid';
-export { default as TopBanner } from './TopBanner/TopBanner';
-export { default as ConfirmDeleteModal } from './ConfirmDeleteModal/ConfirmDeleteModal';
-export { default as NewTemplateButton } from './NewTemplateButton/NewTemplateButton';
-export { default as CategoryListSection } from './CategoryListSection/CategoryListSection';
-export { default as CategoryListSectionSkeleton } from './CategoryListSection/CategoryListSectionSkeleton';
-export { default as TagListSection } from './TagListSection/TagListSection';
-export { default as TagListSectionSkeleton } from './TagListSection/TagListSectionSkeleton';
-export { default as TemplateListSection } from './TemplateListSection/TemplateListSection';
-export { default as TemplateListSectionLoading } from './TemplateListSection/TemplateListSectionLoading';
-export { default as TemplateDeleteSelection } from './TemplateDeleteSelection/TemplateDeleteSelection';
+export { default as CategoryEditModal } from '@/pages/MyTemplatesPage/components/CategoryEditModal/CategoryEditModal';
+export { default as CategoryFilterMenu } from '@/pages/MyTemplatesPage/components/CategoryFilterMenu/CategoryFilterMenu';
+export { default as TagFilterMenu } from '@/pages/MyTemplatesPage/components/TagFilterMenu/TagFilterMenu';
+export { default as TemplateGrid } from '@/pages/MyTemplatesPage/components/TemplateGrid/TemplateGrid';
+export { default as TopBanner } from '@/pages/MyTemplatesPage/components/TopBanner/TopBanner';
+export { default as ConfirmDeleteModal } from '@/pages/MyTemplatesPage/components/ConfirmDeleteModal/ConfirmDeleteModal';
+export { default as NewTemplateButton } from '@/pages/MyTemplatesPage/components/NewTemplateButton/NewTemplateButton';
+export { default as CategoryListSection } from '@/pages/MyTemplatesPage/components/CategoryListSection/CategoryListSection';
+export { default as CategoryListSectionSkeleton } from '@/pages/MyTemplatesPage/components/CategoryListSection/CategoryListSectionSkeleton';
+export { default as TagListSection } from '@/pages/MyTemplatesPage/components/TagListSection/TagListSection';
+export { default as TagListSectionSkeleton } from '@/pages/MyTemplatesPage/components/TagListSection/TagListSectionSkeleton';
+export { default as TemplateListSection } from '@/pages/MyTemplatesPage/components/TemplateListSection/TemplateListSection';
+export { default as TemplateListSectionLoading } from '@/pages/MyTemplatesPage/components/TemplateListSection/TemplateListSectionLoading';
+export { default as TemplateDeleteSelection } from '@/pages/MyTemplatesPage/components/TemplateDeleteSelection/TemplateDeleteSelection';
